Вынесла комментарии в отдельную структуру
