+++ src/components/film-popup.js
@@ -1,69 +1,6 @@
 import AbstractSmartComponent from "./abstract-smart-component";
-import {EMOGIES, TypesButton} from "../consts";
-import {formatReleaseDate, formatDateOfComment} from "../utils/common";
-
-const createCommentsMarkup = (comments) => {
-  return comments
-    .map((comment) => {
-      const commentDate = formatDateOfComment(comment.date);
-
-      return (
-        `<li class="film-details__comment">
-            <span class="film-details__comment-emoji">
-              <img src="./images/emoji/${comment.emoji}.png" width="55" height="55" alt="emoji-angry">
-            </span>
-            <div>
-              <p class="film-details__comment-text">${comment.message}</p>
-              <p class="film-details__comment-info">
-                <span class="film-details__comment-author">${comment.userName}</span>
-                <span class="film-details__comment-day">${commentDate}</span>
-                <button class="film-details__comment-delete">Delete</button>
-              </p>
-            </div>
-          </li>
-        `
-      );
-    }).join(`\n`);
-};
-
-const createInstallEmojiMarkup = (emogies, chosenEmoji) => {
-  return emogies
-    .map((emoji) => {
-      return (
-        `<input
-          class="film-details__emoji-item visually-hidden"
-          name="comment-emoji"
-          type="radio"
-          id="emoji-${emoji}"
-          value="${emoji}"
-          ${chosenEmoji === emoji ? `checked` : ``}
-          >
-                <label class="film-details__emoji-label" for="emoji-${emoji}">
-              <img src="./images/emoji/${emoji}.png" width="30" height="30" alt="emoji">
-            </label>`
-      );
-    }).join(`\n`);
-};
-
-const createNewCommentMarkup = (chosenEmoji) => {
-  const installingEmojiMarkup = createInstallEmojiMarkup(EMOGIES, chosenEmoji);
-
-  return (
-    `<div class="film-details__new-comment">
-          <div for="add-emoji" class="film-details__add-emoji-label">
-            ${chosenEmoji !== null ? `<img src="images/emoji/${chosenEmoji}.png" width="55" height="55" alt="emoji-${chosenEmoji}">` : ``}
-          </div>
-
-          <label class="film-details__comment-label">
-            <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-          </label>
-
-          <div class="film-details__emoji-list">
-            ${installingEmojiMarkup}
-          </div>
-        </div>`
-  );
-};
+import {TypesButton} from "../consts";
+import {formatReleaseDate} from "../utils/common";
 
 const createCheckboxMarkup = (name, isChecked) => {
   const setDescriptionButton = () => {
@@ -92,7 +29,7 @@ const createCheckboxMarkup = (name, isChecked) => {
   );
 };
 
-const createFilmPopupTemplate = (film, options = {}) => {
+const createFilmPopupTemplate = (film) => {
   const {
     name,
     poster,
@@ -110,11 +47,7 @@ const createFilmPopupTemplate = (film, options = {}) => {
     comments
   } = film;
 
-  const {chosenEmoji} = options;
-
   const genreTitle = genres.length > 1 ? `Genres` : `Genre`;
-  const commentsMarkup = createCommentsMarkup(comments);
-  const newCommentMarkup = createNewCommentMarkup(chosenEmoji);
 
   const releaseDate = formatReleaseDate(date);
 
@@ -197,16 +130,7 @@ const createFilmPopupTemplate = (film, options = {}) => {
     </div>
 
     <div class="form-details__bottom-container">
-      <section class="film-details__comments-wrap">
-        <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${comments.length}</span></h3>
-
-        <ul class="film-details__comments-list">
-         ${commentsMarkup}
-        </ul>
-
-        ${newCommentMarkup}
-        </div>
-      </section>
+    
     </div>
   </form>
 </section>`
@@ -217,25 +141,19 @@ export default class FilmPopup extends AbstractSmartComponent {
   constructor(film) {
     super();
     this._film = film;
-    this._chosenEmoji = null;
 
     this._closeHandler = null;
     this._addWatchlistCheckboxChangeHandler = null;
     this._watchedCheckboxChangeHandler = null;
     this._favoriteCheckboxChangeHandler = null;
-
-    this._setEmojiRadioChangeHandler();
   }
 
   getTemplate() {
-    return createFilmPopupTemplate(this._film, {
-      chosenEmoji: this._chosenEmoji,
-    });
+    return createFilmPopupTemplate(this._film);
   }
 
   recoveryListeners() {
     this.setCloseButtonClickHandler(this._closeHandler);
-    this._setEmojiRadioChangeHandler();
     this.setAddWatchlistCheckboxChangeHandler(this._addWatchlistCheckboxChangeHandler);
     this.setWatchedCheckboxChangeHandler(this._watchedCheckboxChangeHandler);
     this.setFavoriteCheckboxChangeHandler(this._favoriteCheckboxChangeHandler);
@@ -246,8 +164,6 @@ export default class FilmPopup extends AbstractSmartComponent {
   }
 
   reset() {
-    this._chosenEmoji = null;
-
     this.rerender();
   }
 
@@ -278,17 +194,5 @@ export default class FilmPopup extends AbstractSmartComponent {
       .addEventListener(`change`, handler);
 
     this._favoriteCheckboxChangeHandler = handler;
-  }
-
-  _setEmojiRadioChangeHandler() {
-    const emogies = this.getElement().querySelectorAll(`.film-details__emoji-item`);
-
-    emogies.forEach((emoji) => {
-      emoji.addEventListener(`change`, (evt) => {
-        this._chosenEmoji = evt.target.value;
-
-        this.rerender();
-      });
-    });
   }
 }
